staking(validator_manager): add `update_validator`
